Refactored edit component
